requirements.txt
allure_python_commons==2.8.34
jsonschema==3.2.0
psycopg2==2.9.3
pycryptodome==3.12.0
pymysql==1.0.2
pytest==6.2.2
PyYAML==6.0
requests==2.24.0
selenium==3.141.0
ujson==5.1.0
xlrd==1.2.0
xlwt==1.3.0
allure-pytest==2.8.34
allure-python-commons==2.8.34
pytest==6.2.2
pytest-dependency==0.5.1
pytest-repeat==0.9.1
pytest-rerunfailures==9.1.1